Refactor osism.tasks.conductor.sonic into modular sonic subdirectory
Break down the large 1700+ line sonic.py module into smaller, more maintainable
files organized in a sonic/ subdirectory:

- sonic/constants.py: Constants, mappings, and configuration defaults
- sonic/device.py: Device-related helper functions (platform, hostname, MAC)
- sonic/interface.py: Interface conversion, port detection, and speed mapping
- sonic/bgp.py: BGP neighbor detection and AS number calculation
- sonic/config_generator.py: Main SONiC configuration generation logic
- sonic/exporter.py: Configuration export functions (NetBox and file)
- sonic/sync.py: Main synchronization function
- sonic/__init__.py: Package initialization with public exports

The original sonic.py is now a compatibility layer that re-exports all functions
from the new modular structure, ensuring backward compatibility.

AI-assisted: Claude Code
