+++ osism/tasks/conductor/sonic/__init__.py
@@ -1,0 +1,14 @@
+# SPDX-License-Identifier: Apache-2.0
+
+"""SONiC configuration management package."""
+
+from .config_generator import generate_sonic_config
+from .exporter import save_config_to_netbox, export_config_to_file
+from .sync import sync_sonic
+
+__all__ = [
+    "generate_sonic_config",
+    "save_config_to_netbox",
+    "export_config_to_file",
+    "sync_sonic",
+]
